removed old tensor implementation
