+++ cpp_src/CMakeLists.txt
@@ -17,18 +17,18 @@ SET(CUDA_PROPAGATE_HOST_FLAGS OFF)
 add_compile_definitions(CUDA=1)
 
 # Link runTests with what we want to test and the GTest and pthread library
-CUDA_ADD_EXECUTABLE(runTests test/tests.cpp utils/signal_handler.cpp gpu/dark.cu gpu/tensor.cu)
+# CUDA_ADD_EXECUTABLE(runTests test/tests.cpp utils/signal_handler.cpp gpu/dark.cu gpu/tensor.cu)
 
-target_include_directories(runTests PUBLIC
-    ${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES}
-    ${CUDA_INCLUDE_DIRS}
-    "../3rdparty"
-    ${GTEST_INCLUDE_DIRS}
-    "."
-)
-TARGET_LINK_LIBRARIES(runTests ${CUDA_LIBRARIES} ${GTEST_LIBRARIES} pthread)
+# target_include_directories(runTests PUBLIC
+#     ${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES}
+#     ${CUDA_INCLUDE_DIRS}
+#     "../3rdparty"
+#     ${GTEST_INCLUDE_DIRS}
+#     "."
+# )
+# TARGET_LINK_LIBRARIES(runTests ${CUDA_LIBRARIES} ${GTEST_LIBRARIES} pthread)
 
-CUDA_ADD_EXECUTABLE(runTests2 test/tests2.cpp)
+CUDA_ADD_EXECUTABLE(runTests2 test/tests2.cpp tensor/tensor_cpu.cpp types/enum.cpp)
 
 target_include_directories(runTests2 PUBLIC
     ${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES}
