--- cpp_src/tensor/tensor.cpp
@@ -1,60 +1,0 @@
-#include "tensor/tensor.hpp"
-
-namespace darknet
-{
-namespace tensor
-{
-
-
-
-    // Tensor(TensorShape shape)
-    //     : shape(shape)
-    // {
-
-    // }
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::CPU>::value>::type>
-    // Tensor(std::initializer_list<int> shape)
-    //     : shape(shape)
-    // {
-
-    // }
-
-    // // GPU CONSTRUCTORS
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::GPU>::value>::type>
-    // Tensor(cuda::device_t _device)
-    //     : _device(_device), shape({})
-    // {
-
-    // }
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::GPU>::value>::type>
-    // Tensor(cuda::device_t _device, TensorShape shape)
-    //     : _device(_device), shape(shape)
-    // {
-
-    // }
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::GPU>::value>::type>
-    // Tensor(cuda::device_t _device, std::initializer_list<int> shape)
-    //     : _device(_device), shape(shape)
-    // {
-
-    // }
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::GPU>::value>::type>
-    // Tensor()
-    //     : _device(cuda::device::default_device_id), shape({})
-    // {
-
-    // }
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::GPU>::value>::type>
-    // Tensor(TensorShape shape)
-    //     : _device(cuda::device::default_device_id), shape(shape)
-    // {
-
-    // }
-    // // template<DeviceType U = device, std::enable_if<std::is_same<U, DeviceType::GPU>::value>::type>
-    // Tensor(std::initializer_list<int> shape)
-    //     : _device(cuda::device::default_device_id), shape(shape)
-    // {
-
-    // }
-} // namespace tensor
-} // namespace darknet
